libstagefright: Support PCM channel mask update
For PCM offloaded content that has a different
channel count when compared to what is
specified in the header, playback fails.
Audio HAL expects the updated channel mask
to be sent to it. Fix this by overwriting
the channel mask retrieved from the header
of the file format with the channel mask
retrieved based on the decoders output

Change-Id: I363ab5a4cf5b6d13ef52fede25a644661b11d834
CRs-Fixed: 825717