Fix issues crashing with Fatal signal when calling getFormat
when playing multimedia file, GenericSource::getTrackInfo run in the
NuPlayerDriver thread, and GenericSource::getFormatMeta run in the
Generic thread. they would access the critical resource which is
mMetaData in IMediaSource. There is no lock to protect mMetaData.
Occasionally, SIGSEGV occurs when mMetaData point to a new object.

Instead getting trackInfo in Generic thread.

Change-Id: Ib648edf89aad34f8bd1cbaa3a51febf57733ec9b
Signed-off-by: qinzhichao <qinzhichao@xiaomi.com>