stagefright: omx: read a property to replace invalid dataspace
It has been observed that the dataspace value is invalid for 2160p video
recording on the Sony Xperia X family of devices. The issue is that the camera HAL
is refusing the passed dataspace with value 0x10c60000.

After checking, the dataspace passed seems to really be invalid, cause
unknown.

The solution in this commit is a property that enables the detection and
replacement of that dataspace. The value is then overriden with an
acceptable value, which is the same dataspace used for 1080p recording.

Change-Id: I11a09f5243a1a649807626ff28d80cb1bb3bf0cb